requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-jose
pydantic
psycopg2-binary
SQLAlchemy
streamlit
streamlit-authenticator
requests
psycopg2-binary
sqlalchemy[asyncio]
asyncpg
python-jose
python-dotenv
pydantic
typing-extensions
langchain
langchain-core
google-generativeai
langchain-community
requests
langgraph
wikipedia
google-search-results
Sentence-transformers
pgvector
passlib
bcrypt